requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.1
datasets>=2.15.0